requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python_dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.12.0
snowflake_id>=0.0.2
crypto>=1.4.1
pyOpenSSL>=23.1.0
pycryptodome>=3.17